Vision/CVR/oscar_requirements.txt
numpy
opencv-python==4.5.5.64
pandas
Pillow==9.0.1
pytorch-lightning==1.6.0
PyYAML
scikit-learn
scipy
tensorboard==2.8.0
timm
